# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.OpenAIAudioTranscriptionResponse:
    properties:
      duration:
        example: 12.34
        type: number
      language:
        example: en
        type: string
      segments:
        items:
          $ref: '#/definitions/docs.OpenAIWhisperSegment'
        type: array
      task:
        example: transcribe
        type: string
      text:
        example: Hello world
        type: string
    type: object
...
  docs.OpenAIError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  docs.OpenAIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/docs.OpenAIError'
    type: object
...
  docs.OpenAIImageData:
    properties:
      b64_json:
        example: base64-image
        type: string
      revised_prompt:
        example: A cute cat
        type: string
      url:
        example: https://example.com/image.png
        type: string
    type: object
  docs.OpenAIImageGenerationRequest:
    properties:
      model:
        example: gpt-image-1
        type: string
      "n":
        example: 1
        type: integer
      prompt:
        example: A cute cat
        type: string
      quality:
        example: standard
        type: string
      response_format:
        example: url
        type: string
      size:
        example: 1024x1024
        type: string
      style:
        example: vivid
        type: string
      user:
        example: user-123
        type: string
    type: object
  docs.OpenAIImageResponse:
    properties:
      created:
        example: 1700000000
        type: integer
      data:
        items:
          $ref: '#/definitions/docs.OpenAIImageData'
        type: array
    type: object
...
  docs.OpenAIModerationCategories:
    properties:
      harassment:
        example: false
        type: boolean
      harassment_threatening:
        example: false
        type: boolean
      hate:
        example: false
        type: boolean
      hate_threatening:
        example: false
        type: boolean
      illicit:
        example: false
        type: boolean
      illicit_violent:
        example: false
        type: boolean
      self_harm:
        example: false
        type: boolean
      self_harm_instructions:
        example: false
        type: boolean
      self_harm_intent:
        example: false
        type: boolean
      sexual:
        example: false
        type: boolean
      sexual_minors:
        example: false
        type: boolean
      violence:
        example: false
        type: boolean
      violence_graphic:
        example: false
        type: boolean
    type: object
  docs.OpenAIModerationCategoryScores:
    properties:
      harassment:
        example: 0
        type: number
      harassment_threatening:
        example: 0
        type: number
      hate:
        example: 0
        type: number
      hate_threatening:
        example: 0
        type: number
      illicit:
        example: 0
        type: number
      illicit_violent:
        example: 0
        type: number
      self_harm:
        example: 0
        type: number
      self_harm_instructions:
        example: 0
        type: number
      self_harm_intent:
        example: 0
        type: number
      sexual:
        example: 0
        type: number
      sexual_minors:
        example: 0
        type: number
      violence:
        example: 0
        type: number
      violence_graphic:
        example: 0
        type: number
    type: object
  docs.OpenAIModerationRequest:
    properties:
      input:
        example:
        - '["I want to hurt someone"]'
        items:
          type: string
        type: array
      model:
        example: omni-moderation-latest
        type: string
    type: object
  docs.OpenAIModerationResponse:
    properties:
      id:
        example: modr-123
        type: string
      model:
        example: omni-moderation-latest
        type: string
      results:
        items:
          $ref: '#/definitions/docs.OpenAIModerationResult'
        type: array
    type: object
  docs.OpenAIModerationResult:
    properties:
      categories:
        $ref: '#/definitions/docs.OpenAIModerationCategories'
      category_scores:
        $ref: '#/definitions/docs.OpenAIModerationCategoryScores'
      flagged:
        example: false
        type: boolean
    type: object
...
  docs.OpenAIWhisperSegment:
    properties:
      avg_logprob:
        example: -0.1
        type: number
      compression_ratio:
        example: 1
        type: number
      end:
        example: 1.2
        type: number
      id:
        example: 0
        type: integer
      no_speech_prob:
        example: 0
        type: number
      seek:
        example: 0
        type: integer
      start:
        example: 0
        type: number
      temperature:
        example: 0
        type: number
      text:
        example: Hello
        type: string
      tokens:
        items:
          type: integer
        type: array
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
...
  /api/v1/admin/tasks:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List async tasks (admin)
      tags:
      - admin
  /api/v1/admin/tasks/{id}:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get async task by ID (admin)
      tags:
      - admin
  /api/v1/admin/tasks/{id}/cancel:
    post:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Cancel async task (admin)
      tags:
      - admin
  /api/v1/admin/tasks/{id}/retry:
    post:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Retry async task (admin)
      tags:
      - admin
...
  /api/v1/admin/user/tasks:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List user tasks (admin)
      tags:
      - admin
  /api/v1/admin/user/tasks/{id}:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get user task by task ID (admin)
      tags:
      - admin
...
  /api/v1/public/audio/transcriptions:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Audio file
        in: formData
        name: file
        required: true
        type: file
      - description: Model ID (e.g., whisper-1)
        in: formData
        name: model
        required: true
        type: string
      - description: Language (ISO-639-1)
        in: formData
        name: language
        type: string
      - description: Prompt
        in: formData
        name: prompt
        type: string
      - description: Response format (json, text, srt, vtt, verbose_json)
        in: formData
        name: response_format
        type: string
      - description: Sampling temperature
        in: formData
        name: temperature
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.OpenAIAudioTranscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: Audio transcriptions (OpenAI compatible)
      tags:
      - public
  /api/v1/public/audio/translations:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Audio file
        in: formData
        name: file
        required: true
        type: file
      - description: Model ID (e.g., whisper-1)
        in: formData
        name: model
        required: true
        type: string
      - description: Prompt
        in: formData
        name: prompt
        type: string
      - description: Response format (json, text, srt, vtt, verbose_json)
        in: formData
        name: response_format
        type: string
      - description: Sampling temperature
        in: formData
        name: temperature
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.OpenAIAudioTranscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: Audio translations (OpenAI compatible)
      tags:
      - public
...
  /api/v1/public/images/generations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Image generation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/docs.OpenAIImageGenerationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.OpenAIImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: Image generations (OpenAI compatible)
      tags:
      - public
...
  /api/v1/public/moderations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Moderations request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/docs.OpenAIModerationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.OpenAIModerationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: Moderations (OpenAI compatible)
      tags:
      - public
...
  /api/v1/public/user/tasks:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List current user tasks
      tags:
      - user
  /api/v1/public/user/tasks/{id}:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get current user task by task ID
      tags:
      - user
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"